feat!: replace tracing_enabled with tracing_provider (#156)
## Description
https://github.com/justtrackio/gosoline/pull/1170 will replace the
`tracing.enabled` setting with the `tracing.provider` setting. Therefore
we need an update to be able to configure tracing in the future as well.

## Motivation and Context
Without this change, we can't configure tracing anymore.

## Breaking Changes
This is a breaking change. If you had `tracing_enabled = true` in the
past, you have to replace this with `tracing_provider = "xray"`. If it
was set to false, you have to use `local` (recommended) or `noop` as
your tracing provider.

## How Has This Been Tested?
- [ ] ~I have updated at least one of the `examples/*` to demonstrate
and validate my change(s)~ There are no examples in this project
- [ ] ~I have tested and validated these changes using one or more of
the provided `examples/*` projects~
- [x] I have executed `pre-commit run -a` on my pull request
- [x] I am updating our internal projects to this version